Update TT_New_Context documentation
In commit 531d463aed365b

  [truetype] Allocate TT_ExecContext in TT_Size instead of TT_Driver.

the `TT_ExecContext` was moved from being on the driver to being on the size
to make it easier to use FreeType in a multi-threaded environment.  However,
the documentation for `TT_New_Context` was not updated and still reflects
the old behavior and parameter list.

This change updates `TT_New_Context` documentation to reflect the current
parameters and usage.